backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.0.0
pandas==2.1.4
openai==1.3.7
boto3==1.34.34
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
openpyxl==3.1.2
Pillow==10.1.0
